Wrap the project for the first version release
Introduce linter function for linting all whitespace from the text, this
is an improvement over the previous version that was only stripping \n
(return characters).

The program in its semi-complete version (without printing of objects)
and can be ran from maillinter.py and the objects have to be printed
manually.